pyproject.toml
[project]
name = "agent-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.6.0",
    "crewai-tools>=0.25.5",
    "crewai[tools]>=0.86.0",
    "flash-attn==2.7.4.post1",
    "google-adk>=0.1.0",
    "jupyter>=1.1.1",
    "litellm>=1.66.1",
    "loguru>=0.7.3",
    "qwen-vl-utils>=0.0.11",
    "sglang[all]>=0.4.6",
    "vllm>=0.8.3",
    "whisperx>=3.3.2",
]

[tool.uv.sources]
google-adk = { git = "https://github.com/google/adk-python.git", rev = "main" }
